+++ turnus.model/src/turnus/model/dataflow/impl/TypeDoubleImpl.java
@@ -74,7 +74,7 @@ public class TypeDoubleImpl extends MinimalEObjectImpl.Container implements Type
 	 * @generated
 	 * @ordered
 	 */
-	protected static final int BITS_EDEFAULT = 0;
+	protected static final long BITS_EDEFAULT = 0L;
 
 	/**
 	 * The default value of the '{@link #getSize() <em>Size</em>}' attribute.
@@ -117,7 +117,7 @@ public class TypeDoubleImpl extends MinimalEObjectImpl.Container implements Type
 	}
 
 	@Override
-	public int getBits() {
+	public long getBits() {
 		return 32 * getSize();
 	}
