+++ turnus.model/src/turnus/model/dataflow/impl/TypeIntImpl.java
@@ -77,7 +77,7 @@ public class TypeIntImpl extends MinimalEObjectImpl.Container implements TypeInt
 	 * @generated
 	 * @ordered
 	 */
-	protected static final int BITS_EDEFAULT = 0;
+	protected static final long BITS_EDEFAULT = 0L;
 
 	/**
 	 * The default value of the '{@link #getSize() <em>Size</em>}' attribute.
@@ -124,7 +124,7 @@ public class TypeIntImpl extends MinimalEObjectImpl.Container implements TypeInt
 	}
 
 	@Override
-	public int getBits() {
+	public long getBits() {
 		return getSize();
 	}
